Make sure legacy tests will be executed
