+++ test/unit/legacy/test_error_code_format.py
