+++ test/unit/legacy/test_exa_error.py
